requirements.txt
dj-database-url==0.5.0
Django==3.1.14
django-cors-headers==3.2.0
django-picklefield==3.0.1
django-q==1.3.3
django-request-logging==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.11.2
fyle==v0.28.0
fyle-rest-auth==1.1.0
psycopg2-binary==2.8.4
pylint==2.7.4
requests==2.25.0
pylint-django==2.3.0
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-mock==3.6.1
